Seperated MidiInputReceiver from PlaybackController
